requirements.txt
cohere
chromadb
python-dotenv
